+++ lib/file_system/include/manager_files_cli.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "file.hpp"
 
 namespace fs {
